chore(refactor): consolidate some of the components and use existing styles
