+++ ui/src/app/Onboarding.tsx
@@ -11,7 +11,7 @@ import {
   UsersIcon
 } from '@heroicons/react/24/outline';
 import { useDispatch } from 'react-redux';
-import Button from '~/components/forms/buttons/Button';
+import { Button } from '~/components/Button';
 import { Icon } from '~/types/Icon';
 import { cls } from '~/utils/helpers';
 import { useNavigate } from 'react-router-dom';
