+++ ui/src/app/console/Console.tsx
@@ -14,9 +14,9 @@ import { selectCurrentNamespace } from '~/app/namespaces/namespacesSlice';
 import { selectCurrentRef } from '~/app/refs/refsSlice';
 import { ContextEditor } from '~/components/console/ContextEditor';
 import EmptyState from '~/components/EmptyState';
-import { Button } from '~/components/ui/button';
-import Combobox from '~/components/forms/Combobox';
-import Dropdown from '~/components/forms/Dropdown';
+import { Button } from '~/components/Button';
+import Combobox from '~/components/Combobox';
+import Dropdown from '~/components/Dropdown';
 import Input from '~/components/forms/Input';
 import { evaluateURL, evaluateV2 } from '~/data/api';
 import { useError } from '~/data/hooks/error';
@@ -268,13 +268,12 @@ export default function Console() {
                               aria-label="New Entity ID"
                               title="New Entity ID"
                               variant="ghost"
-                              size="icon"
                               onClick={(e) => {
                                 e.preventDefault();
                                 formik.setFieldValue('entityId', uuidv4());
                               }}
                             >
-                              <RefreshCwIcon className="text-gray-400" />
+                              <RefreshCwIcon className="h-4 w-4 text-gray-400" />
                             </Button>
                           </div>
                         </div>
@@ -318,7 +317,7 @@ export default function Console() {
                           ]}
                         />
                         <Button
-                          variant="default"
+                          variant="primary"
                           type="submit"
                           disabled={!(formik.dirty && formik.isValid)}
                         >
