+++ ui/src/app/flags/Flag.tsx
@@ -10,7 +10,7 @@ import {
   selectNamespaces
 } from '~/app/namespaces/namespacesSlice';
 import FlagForm from '~/components/flags/forms/FlagForm';
-import Dropdown from '~/components/forms/Dropdown';
+import Dropdown from '~/components/Dropdown';
 import Loading from '~/components/Loading';
 import Modal from '~/components/Modal';
 import MoreInfo from '~/components/MoreInfo';
