+++ ui/src/app/flags/Flags.tsx
@@ -2,8 +2,7 @@ import { useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { selectReadonly } from '~/app/meta/metaSlice';
 import { selectCurrentNamespace } from '~/app/namespaces/namespacesSlice';
-import { Plus } from 'lucide-react';
-import { Button } from '~/components/ui/button';
+import { ButtonWithPlus } from '~/components/Button';
 import FlagTable from '~/components/flags/FlagTable';
 import { PageHeader } from '~/components/ui/page';
 
@@ -18,10 +17,13 @@ export default function Flags() {
   return (
     <>
       <PageHeader title="Flags">
-        <Button onClick={() => navigate(`${path}/new`)} disabled={readOnly}>
-          <Plus />
+        <ButtonWithPlus
+          variant="primary"
+          onClick={() => navigate(`${path}/new`)}
+          disabled={readOnly}
+        >
           New Flag
-        </Button>
+        </ButtonWithPlus>
       </PageHeader>
       <div className="flex flex-col gap-1 space-y-2 py-2">
         <FlagTable namespace={namespace} />
