+++ ui/src/app/tokens/Tokens.tsx
@@ -5,7 +5,7 @@ import {
   useListTokensQuery
 } from '~/app/tokens/tokensApi';
 import EmptyState from '~/components/EmptyState';
-import { ButtonWithPlus } from '~/components/forms/buttons/Button';
+import { ButtonWithPlus } from '~/components/Button';
 import Loading from '~/components/Loading';
 import Modal from '~/components/Modal';
 import DeletePanel from '~/components/panels/DeletePanel';
