+++ ui/src/components/flags/FlagTable.tsx
@@ -19,10 +19,10 @@ import {
   setSorting,
   useListFlagsQuery
 } from '~/app/flags/flagsApi';
-import { cn } from '~/lib/utils';
-import { Badge } from '~/components/ui/badge';
+import { cls } from '~/utils/helpers';
+import { Badge } from '~/components/Badge';
 import { formatDistanceToNowStrict, parseISO } from 'date-fns';
-import { Search } from '~/components/ui/search';
+import Searchbox from '~/components/Searchbox';
 import { DataTableViewOptions } from '~/components/ui/table-view-options';
 import { VariableIcon, ToggleLeftIcon } from 'lucide-react';
 import { useError } from '~/data/hooks/error';
@@ -176,11 +176,11 @@ export default function FlagTable(props: FlagTableProps) {
   return (
     <>
       <div className="flex items-center justify-between">
-        <div className="flex flex-1 items-center justify-between space-x-2">
-          <Search
+        <div className="flex flex-1 items-center justify-between">
+          <Searchbox
             value={filter ?? ''}
             onChange={setFilter}
-            className="h-8 w-[150px] flex-grow text-xs lg:w-[250px]"
+            className="text-xs"
           />
           <DataTableViewOptions table={table} />
         </div>
@@ -208,7 +208,7 @@ export default function FlagTable(props: FlagTableProps) {
           <button
             role="link"
             key={row.id}
-            className={cn(
+            className={cls(
               'flex flex-col items-start gap-2 rounded-lg border p-3 text-left text-sm transition-all hover:bg-accent'
             )}
             onClick={() => navigate(`${path}/${item.key}`)}
