+++ ui/src/components/namespaces/NamespaceForm.tsx
@@ -7,7 +7,7 @@ import {
   useCreateNamespaceMutation,
   useUpdateNamespaceMutation
 } from '~/app/namespaces/namespacesSlice';
-import Button from '~/components/forms/buttons/Button';
+import { Button } from '~/components/Button';
 import Input from '~/components/forms/Input';
 import Loading from '~/components/Loading';
 import MoreInfo from '~/components/MoreInfo';
