+++ ui/src/components/rollouts/Rollout.tsx
@@ -3,7 +3,7 @@ import { IFlag } from '~/types/Flag';
 import { IRollout, rolloutTypeToLabel } from '~/types/Rollout';
 import { ISegment } from '~/types/Segment';
 import { cls } from '~/utils/helpers';
-import Dropdown from '~/components/forms/Dropdown';
+import Dropdown from '~/components/Dropdown';
 import QuickEditRolloutForm from './forms/QuickEditRolloutForm';
 
 type RolloutProps = {
