+++ ui/src/components/segments/SegmentTable.tsx
@@ -17,10 +17,10 @@ import {
   SegmentMatchType,
   segmentMatchTypeToLabel
 } from '~/types/Segment';
-import { cn } from '~/lib/utils';
-import { Badge } from '~/components/ui/badge';
+import { cls } from '~/utils/helpers';
+import { Badge } from '~/components/Badge';
 import { formatDistanceToNowStrict, parseISO } from 'date-fns';
-import { Search } from '~/components/ui/search';
+import Searchbox from '~/components/Searchbox';
 import { DataTableViewOptions } from '~/components/ui/table-view-options';
 import { Link, useNavigate } from 'react-router-dom';
 import { DataTablePagination } from '~/components/ui/table-pagination';
@@ -167,11 +167,11 @@ export default function SegmentTable(props: SegmentTableProps) {
   return (
     <>
       <div className="flex items-center justify-between">
-        <div className="flex flex-1 items-center justify-between space-x-2">
-          <Search
+        <div className="flex flex-1 items-center justify-between">
+          <Searchbox
             value={filter ?? ''}
             onChange={setFilter}
-            className="h-8 w-[150px] flex-grow text-xs lg:w-[250px]"
+            className="text-xs"
           />
           <DataTableViewOptions table={table} />
         </div>
@@ -199,7 +199,7 @@ export default function SegmentTable(props: SegmentTableProps) {
           <button
             role="link"
             key={row.id}
-            className={cn(
+            className={cls(
               'flex flex-col items-start gap-2 rounded-lg border p-3 text-left text-sm transition-all hover:bg-accent'
             )}
             onClick={() => navigate(`${path}/${item.key}`)}
