+++ ui/src/components/variants/forms/VariantForm.tsx
@@ -9,7 +9,7 @@ import {
   useUpdateVariantMutation
 } from '~/app/flags/flagsApi';
 import { selectCurrentNamespace } from '~/app/namespaces/namespacesSlice';
-import Button from '~/components/forms/buttons/Button';
+import { Button } from '~/components/Button';
 import Input from '~/components/forms/Input';
 import TextArea from '~/components/forms/TextArea';
 import Loading from '~/components/Loading';
